+++ src/components/HorizontalScrollingTextBar/HorizontalScrollingTextBar.css
